+++ apps/warehouse/src/jo-approver/jo-approver.resolver.ts
@@ -5,12 +5,13 @@ import { CreateJoApproverInput } from './dto/create-jo-approver.input';
 import { UpdateJoApproverInput } from './dto/update-jo-approver.input';
 import { WarehouseRemoveResponse } from '../__common__/classes';
 import { GqlAuthGuard } from '../__auth__/guards/gql-auth.guard';
-import { UseGuards } from '@nestjs/common';
+import { ForbiddenException, UseGuards } from '@nestjs/common';
 import { AuthUser } from '../__common__/auth-user.entity';
 import { CurrentAuthUser } from '../__auth__/current-auth-user.decorator';
 import { Employee } from '../__employee__/entities/employee.entity';
 import { UpdateJOOrderResponse } from './entities/update-jo-order-response.entity';
 import { UpdateJOOrderInput } from './dto/update-jo-order.input'
+import { isAdmin } from '../__common__/helpers';
 
 @UseGuards(GqlAuthGuard)
 @Resolver(() => JOApprover)
@@ -22,6 +23,11 @@ export class JoApproverResolver {
     @Args('input') createJoApproverInput: CreateJoApproverInput,
     @CurrentAuthUser() authUser: AuthUser
   ) {
+
+    if (!isAdmin(authUser)) {
+      throw new ForbiddenException('Only Admin can create JO Approver')
+    }
+
     this.joApproverService.setAuthUser(authUser)
     return this.joApproverService.create(createJoApproverInput);
   }
@@ -50,12 +56,24 @@ export class JoApproverResolver {
     @Args('input') updateJoApproverInput: UpdateJoApproverInput,
     @CurrentAuthUser() authUser: AuthUser
   ) {
+
+    if (!isAdmin(authUser)) {
+      throw new ForbiddenException('Only Admin can update JO Approver')
+    }
+
     this.joApproverService.setAuthUser(authUser)
     return this.joApproverService.update(id, updateJoApproverInput);
   }
 
   @Mutation(() => UpdateJOOrderResponse)
-  async updateJOApproverOrder(@Args('inputs', { type: () => [UpdateJOOrderInput] }) inputs: UpdateJOOrderInput[]) {
+  async updateJOApproverOrder(
+    @Args('inputs', { type: () => [UpdateJOOrderInput] }) inputs: UpdateJOOrderInput[],
+    @CurrentAuthUser() authUser: AuthUser
+  ) {
+
+    if (!isAdmin(authUser)) {
+      throw new ForbiddenException('Only Admin can update JO Approver Order')
+    }
 
     return await this.joApproverService.updateManyOrders(inputs);
 
@@ -66,6 +84,11 @@ export class JoApproverResolver {
     @Args('id') id: string,
     @CurrentAuthUser() authUser: AuthUser
   ) {
+
+    if (!isAdmin(authUser)) {
+      throw new ForbiddenException('Only Admin can remove JO Approver')
+    }
+
     this.joApproverService.setAuthUser(authUser)
     return this.joApproverService.remove(id);
   }
