+++ apps/warehouse/src/spr-approver/spr-approver.resolver.ts
@@ -5,12 +5,13 @@ import { CreateSprApproverInput } from './dto/create-spr-approver.input';
 import { UpdateSprApproverInput } from './dto/update-spr-approver.input';
 import { WarehouseRemoveResponse } from '../__common__/classes';
 import { GqlAuthGuard } from '../__auth__/guards/gql-auth.guard';
-import { UseGuards } from '@nestjs/common';
+import { ForbiddenException, UseGuards } from '@nestjs/common';
 import { AuthUser } from '../__common__/auth-user.entity';
 import { CurrentAuthUser } from '../__auth__/current-auth-user.decorator';
 import { Employee } from '../__employee__/entities/employee.entity';
 import { UpdateSPROrderResponse } from './entities/update-spr-order-response.entity';
 import { UpdateSPROrderInput } from './dto/update-spr-order.input'
+import { isAdmin } from '../__common__/helpers';
 
 @UseGuards(GqlAuthGuard)
 @Resolver(() => SPRApprover)
@@ -22,6 +23,11 @@ export class SprApproverResolver {
     @Args('input') createSprApproverInput: CreateSprApproverInput,
     @CurrentAuthUser() authUser: AuthUser
   ) {
+
+    if (!isAdmin(authUser)) {
+      throw new ForbiddenException('Only Admin can create SPR Approver')
+    }
+
     this.sprApproverService.setAuthUser(authUser)
     return this.sprApproverService.create(createSprApproverInput);
   }
@@ -50,12 +56,24 @@ export class SprApproverResolver {
     @Args('input') updateSprApproverInput: UpdateSprApproverInput,
     @CurrentAuthUser() authUser: AuthUser
   ) {
+
+    if (!isAdmin(authUser)) {
+      throw new ForbiddenException('Only Admin can update SPR Approver')
+    }
+
     this.sprApproverService.setAuthUser(authUser)
     return this.sprApproverService.update(id, updateSprApproverInput);
   }
 
   @Mutation(() => UpdateSPROrderResponse)
-  async updateSPRApproverOrder(@Args('inputs', { type: () => [UpdateSPROrderInput] }) inputs: UpdateSPROrderInput[]) {
+  async updateSPRApproverOrder(
+    @Args('inputs', { type: () => [UpdateSPROrderInput] }) inputs: UpdateSPROrderInput[],
+    @CurrentAuthUser() authUser: AuthUser
+  ) {
+
+    if (!isAdmin(authUser)) {
+      throw new ForbiddenException('Only Admin can update SPR Approver Order')
+    }
 
     return await this.sprApproverService.updateManyOrders(inputs);
 
@@ -66,6 +84,11 @@ export class SprApproverResolver {
     @Args('id') id: string,
     @CurrentAuthUser() authUser: AuthUser
   ) {
+
+    if (!isAdmin(authUser)) {
+      throw new ForbiddenException('Only Admin can remove SPR Approver')
+    }
+
     this.sprApproverService.setAuthUser(authUser)
     return this.sprApproverService.remove(id);
   }
